// src/categories/enum/user-category.enum.ts
export enum UserCategory {
  MILITARY = 'Armed Forces Personnel',
  MILITARY_FAMILY = 'Families',
  MINDEF_EMPLOY = 'MINDEF Employees',
  VETERANS = 'Veterans',
  MILITARY_MS = 'Personel Angkatan Tentera',
  VETERANS_MS = 'Veteran',
  MILITARY_FAMILY_MS = 'Keluarga',
  MINDEF_EMPLOY_MS = 'Pegawai MINDEF',
}